customising the admin panel

+++ listing/models.py
@@ -15,7 +15,7 @@ class listing(models.Model):
     description = models.TextField(blank=True)
     price = models.IntegerField()
     bedrooms = models.IntegerField()
-    bathrooms = models.DecimalField(max_digits=2, decimal_places=2)
+    bathrooms = models.IntegerField()
     garage = models.IntegerField(default=0)
     sqft = models.IntegerField()
     is_published = models.BooleanField(default=True)
